+++ lib/veev_utils.dart
@@ -1,3 +1,5 @@
 library veev_utils;
 
 export 'src/classes/cep/cep_class.dart';
+export 'src/classes/conversores/numeros_string.dart';
+export 'src/classes/cpf/cpf_validator_class.dart';