+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.3] - Funções para validar Cpf e extrair numeros de uma String implementadas.
+
+* Validar CPF adicionado.
+* Extrair numeros de uma String.
+
 ## [0.0.2] - Comentarios adicionados.
 
 * Comentarios adicionados.